==> 0_mad_per_participant.Rmd <==
#mad-transform data
```{r}
 
# Calculate the median
median_value <- median(fpafod$`Hr [1/min]_EcgMove4`, na.rm = TRUE)#Hr [1/min]_ecgMove4 #`Hr [1/min]_EcgMove4`
# Subtract the median from each value
centered_values <- fpafod$`Hr [1/min]_EcgMove4` - median_value
# Calculate the MAD
mad_value <- mad(centered_values, na.rm = TRUE)
# Normalize the centered values by dividing by the MAD
normalized_values <- centered_values / mad_value
fpafod$HR_mad <- normalized_values

#throw out values that are more than 3 MAD different from the median
fpafod$HR_mad <- ifelse(fpafod$HR_mad > 5, NA, fpafod$HR_mad)#changed this to value 5 with VP 52! (VP 53-60 had value 3)
fpafod$HR_mad <- ifelse(fpafod$HR_mad < -5, NA, fpafod$HR_mad)

write_csv(fpafod, file = "D:/Datamost/ExperienceMapsDataset.csv")


```
